#877 Move `Node` implementations to .util, clean up remaining refactored packages
