+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/reporting/DiagnosticsReport.java
@@ -16,6 +16,8 @@
 
 package org.dockbox.hartshorn.reporting;
 
+import org.dockbox.hartshorn.util.Node;
+
 public interface DiagnosticsReport {
 
     Node<?> root();
