--- hartshorn-reporting/src/main/java/org/dockbox/hartshorn/reporting/proxy/ProxyDiagnosticsReporter.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019-2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dockbox.hartshorn.reporting.proxy;
-
-import org.dockbox.hartshorn.reporting.ConfigurableDiagnosticsReporter;
-import org.dockbox.hartshorn.reporting.collect.DiagnosticsReportCollector;
-
-public class ProxyDiagnosticsReporter implements ConfigurableDiagnosticsReporter<ProxyReportingConfiguration> {
-
-    private final ProxyReportingConfiguration configuration = new ProxyReportingConfiguration();
-
-    @Override
-    public void report(final DiagnosticsReportCollector collector) {
-        // TODO: Implement
-    }
-
-    @Override
-    public ProxyReportingConfiguration configuration() {
-        return this.configuration;
-    }
-}
