--- hartshorn-reporting/src/main/java/org/dockbox/hartshorn/reporting/proxy/ProxyReportingConfiguration.java
@@ -1,51 +1,0 @@
-/*
- * Copyright 2019-2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dockbox.hartshorn.reporting.proxy;
-
-public class ProxyReportingConfiguration {
-
-    private boolean includeProxyType = true;
-    private boolean includeInterfaces = true;
-    private boolean includeInterceptors = true;
-
-    public boolean includeProxyType() {
-        return this.includeProxyType;
-    }
-
-    public ProxyReportingConfiguration includeProxyType(final boolean includeProxyType) {
-        this.includeProxyType = includeProxyType;
-        return this;
-    }
-
-    public boolean includeInterfaces() {
-        return this.includeInterfaces;
-    }
-
-    public ProxyReportingConfiguration includeInterfaces(final boolean includeInterfaces) {
-        this.includeInterfaces = includeInterfaces;
-        return this;
-    }
-
-    public boolean includeInterceptors() {
-        return this.includeInterceptors;
-    }
-
-    public ProxyReportingConfiguration includeInterceptors(final boolean includeInterceptors) {
-        this.includeInterceptors = includeInterceptors;
-        return this;
-    }
-}
