+++ hartshorn-reporting/src/test/java/test/org/dockbox/hartshorn/reporting/ApplicationReportingTests.java
@@ -22,8 +22,8 @@ import org.dockbox.hartshorn.component.factory.UseFactoryServices;
 import org.dockbox.hartshorn.proxy.UseProxying;
 import org.dockbox.hartshorn.reporting.DiagnosticsReport;
 import org.dockbox.hartshorn.reporting.DiagnosticsReportCollector;
-import org.dockbox.hartshorn.reporting.GroupNode;
-import org.dockbox.hartshorn.reporting.Node;
+import org.dockbox.hartshorn.util.GroupNode;
+import org.dockbox.hartshorn.util.Node;
 import org.dockbox.hartshorn.reporting.Reportable;
 import org.dockbox.hartshorn.reporting.application.ApplicationDiagnosticsReporter;
 import org.dockbox.hartshorn.reporting.component.ComponentDiagnosticsReporter;
